[AAE-5617] rename mock icons file
