+++ src/app/stories/document-list/mock/mock-icons.ts
